Use details/summary for class index
This switches from an ordered list to a single span for classes
without subentries, and details/summary for classes with subentries.

Because hanna is most often used in single libraries, the first
details is left open, as most libraries only have one top level
class, and this will show all entries of that top level class.

Inspired by docs.ruby-lang.org, which uses a similar approach
(though it doesn't open the top level by default, since it is
for all core/stdlib classes).