More mobile improvements
Add word break to the header, to prevent overflow that makes the whole
page horizontally scrollable on mobile.

Display the method list within normal flow of the document on mobile, as
there is not enough horizontal place to display the method list
comfortably (it's too narrow, and overflow is cut off). The additional
horizontal space for the main content is especially useful for plugin
classes that only have a few of uninteresting methods.

Add horizontal scrolling to various sections whose content would
otherwise overflow and cause the whole page to be horizontally
scrollable.

Remove the horizontal margin for code blocks, to be able to display more
characters. On mobile it looks crammed when code blocks are indented,
because there is little horizontal space as it is.